sky_zhiyi/backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
python-docx>=1.1.0
PyPDF2>=3.0.0
